+++ Essentials/src/main/java/com/earth2me/essentials/commands/Commandbalancetop.java
@@ -131,7 +131,7 @@ public class Commandbalancetop extends EssentialsCommand {
 
                         // Checking if player meets the requirements of minimum balance and minimum playtime to be listed in baltop list
                         if ((ess.getSettings().showZeroBaltop() || entry.getValue().getBalance().compareTo(BigDecimal.ZERO) > 0)
-                                && balance.compareTo(BigDecimal.valueOf(ess.getSettings().getBaltopMinBalance())) >= 0 &&
+                                && balance.compareTo(ess.getSettings().getBaltopMinBalance()) >= 0 &&
                                 playTimeSecs > ess.getSettings().getBaltopMinPlaytime()) {
                             newCache.getLines().add(tl("balanceTopLine", pos, entry.getValue().getDisplayName(), NumberUtil.displayCurrency(entry.getValue().getBalance(), ess)));
                         }
